Add files via upload
The design process for the photo opportunity at the Hong Leong Bank launch was a carefully orchestrated fusion of branding elements and the core message. Drawing inspiration from the logos of both Pay Watch and Hong Leong Bank, the photo op was thoughtfully crafted to visually represent the collaborative spirit of the partnership. The focal point of the photo op was the powerful message "Any Day Is Pay Day," which served as a dynamic backdrop, symbolizing the essence of the innovative Earned Wage Access (EWA) solution introduced at the launch.